+++ FuzzyAutocomplete/DVTTextCompletionListWindowController+FuzzyAutocomplete.m
@@ -44,6 +44,10 @@
     [self jr_swizzleMethod: @selector(_updateCurrentDisplayStateForQuickHelp)
                 withMethod: @selector(_fa_updateCurrentDisplayStateForQuickHelp)
                      error: NULL];
+
+    [self jr_swizzleMethod: @selector(_preferredWindowFrameForTextFrame:columnsWidth:titleColumnX:)
+                withMethod: @selector(_fa_preferredWindowFrameForTextFrame:columnsWidth:titleColumnX:)
+                     error: nil];
 }
 
 #pragma mark - overrides
@@ -138,6 +142,13 @@ const char kRowHeightKey;
     }
 }
 
+// We modify titleColumnX with score column's width, so that title column is aligned with typed text
+- (CGRect) _fa_preferredWindowFrameForTextFrame: (CGRect) textFrame columnsWidth: (double *) widths titleColumnX: (double) titleX {
+    NSTableView * tableView = [self valueForKey: @"_completionsTableView"];
+    return [self _fa_preferredWindowFrameForTextFrame: textFrame columnsWidth: widths titleColumnX: titleX + [self _fa_widthForScoreColumn] + tableView.intercellSpacing.width];
+}
+
+
 // We add visual feedback for the matched ranges. Also format the score column.
 - (void) _fa_tableView: (NSTableView *) aTableView
        willDisplayCell: (NSCell *) aCell
@@ -152,6 +163,7 @@ const char kRowHeightKey;
     } else if ([aTableColumn.identifier isEqualToString:@"title"]) {
         id<DVTTextCompletionItem> item = self.session.filteredCompletionsAlpha[rowIndex];
         NSArray * ranges = [self.session fa_matchedRangesForItem: item];
+        NSArray * second = [self.session fa_secondPassMatchedRangesForItem: item];
 
         if (!ranges.count) {
             return;
@@ -164,10 +176,34 @@ const char kRowHeightKey;
                                        toString: item.displayText
                                       addOffset: 0];
 
+        second = [self.session fa_convertRanges: second
+                                     fromString: item.name
+                                       toString: item.displayText
+                                      addOffset: 0];
+
         NSDictionary * attributes = [FATheme cuurrentTheme].listTextAttributesForMatchedRanges;
+        NSDictionary * altAttributes = [FATheme cuurrentTheme].listTextAttributesForSecondPassMatchedRanges;
 
         for (NSValue * val in ranges) {
             [attributed addAttributes: attributes range: [val rangeValue]];
+        }
+        for (NSValue * val in second) {
+            [attributed addAttributes: altAttributes range: [val rangeValue]];
+        }
+
+
+        NSString * prefix = self.session.usefulPrefix;
+        if ([item.name hasPrefix: prefix] && ![item.displayText hasPrefix: prefix]) {
+            NSRange prefixRange = NSMakeRange(0, prefix.length);
+            NSArray * prefixRanges = [self.session fa_convertRanges: @[ [NSValue valueWithRange: prefixRange] ]
+                                                         fromString: item.name
+                                                           toString: item.displayText
+                                                          addOffset: 0];
+            prefixRange = NSMakeRange(0, 0);
+            for (NSValue * val in prefixRanges) {
+                prefixRange = NSUnionRange(prefixRange, [val rangeValue]);
+            }
+            [attributed addAttributes: self._usefulPrefixAttributes range: prefixRange];
         }
 
         [aCell setAttributedStringValue: attributed];