+++ FuzzyAutocomplete/DVTTextCompletionSession+FuzzyAutocomplete.h
@@ -30,6 +30,9 @@
 /// Gets array of ranges matched by current search string in given items name.
 - (NSArray *) fa_matchedRangesForItem: (id<DVTTextCompletionItem>) item;
 
+/// Gets array of ranges matched in second pass by current search string in given items name.
+- (NSArray *) fa_secondPassMatchedRangesForItem: (id<DVTTextCompletionItem>) item;
+
 /// Retrieves a previously calculated autocompletion score for given item.
 - (NSNumber *) fa_scoreForItem: (id<DVTTextCompletionItem>) item;
 
@@ -38,9 +41,11 @@
 ///
 /// Handled cases:
 ///
-/// a) fromString is a substring of toString
+/// a) fromString is a substring of toString - offset ranges
 ///
-/// b) both fromString and toString contain segments ending with colons
+/// b) both fromString and toString contain word segments
+///    try to find ranges within segments (may divide ranges)
+///
 - (NSArray *) fa_convertRanges: (NSArray *) originalRanges
                     fromString: (NSString *) fromString
                       toString: (NSString *) toString