+++ FuzzyAutocomplete/FAItemScoringMethod.h
@@ -24,6 +24,7 @@
 /// Scores a single item based on various parameters.
 - (double) scoreItem: (id<DVTTextCompletionItem>) item
         searchString: (NSString *) query
+         matchedName: (NSString *) matchedName
           matchScore: (double) matchScore
        matchedRanges: (NSArray *) rangesArray
       priorityFactor: (double) priorityFactor;